+++ packages/adapter-interfaces/src/types/index.ts
@@ -1,5 +1,4 @@
 export * from "./auth0";
-export * from "./Application";
 export * from "./Client";
 export * from "./ClientGrant";
 export * from "./Flows";