+++ packages/drizzle/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @authhero/drizzle
 
+## 0.4.5
+
+### Patch Changes
+
+- Updated dependencies [149ab91]
+- Updated dependencies [b0e9595]
+  - @authhero/adapter-interfaces@0.91.0
+
 ## 0.4.4
 
 ### Patch Changes