+++ packages/kysely/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @authhero/kysely-adapter
 
+## 10.45.0
+
+### Minor Changes
+
+- 149ab91: Drop the old application table
+- b0e9595: Add client grants
+
+### Patch Changes
+
+- Updated dependencies [149ab91]
+- Updated dependencies [b0e9595]
+  - @authhero/adapter-interfaces@0.91.0
+
 ## 10.44.0
 
 ### Minor Changes